Basically rewriting it all
I've removed "proto" because it was redundant from what I could tell?

Moved ._access to the lowest metatable so we can always just check there to see if we ever were given access. This is cause like, we could jump through stuff a bunch and remove metatables and things like that so like this it's easy to always fall back.

Fully changed how super works. just self:super now. Was a bit weird to get working but I knew all the info was already there, I just needed to handle it correctly

MONKEYPATCH REMOVED. It's a seperate thing called "extended-types" now, its on github

worked on index and newindex and I think I got all the edge cases. For example, there needed to be a rawset under priv variables as well as a rawget for priv gets, otherwise S.O

Don't keep adding sets or gets as you look down the tables for getters and setters. When I was testing it that wasn't happening but I'm SURE that it should have been happening, because index and newindex fire again, and then k would be set_ key in that case and add another set_ for each successive metatable basically. Don't know why it didn't do that but ya, I made sure it wouldnt

Some of the jank that is in the commented version is fixed but it was only there because of me trying to implement private variables correctly