Final portion of rewrite
Reimplemented is and implement, not much was needed to change.

added a :tostring for Object so it had something to fallback to prevent errors when trying to tostring something and someone hadn't implemented a tostring themselves. Grabs type, so if they havent declared a type it'll also just say "object"

Pretty confident about it and also it's MINE NOW since I've rewrote the whole damn thing xD